[WIP] Proximity match: identify types based on attribute probability
We cannot express all the existing symbols in the world. We don't have them all
and it'd be computably expensive.

Say we have a make and model car, Ford Fusion. We don't know the Fusion symbol,
but we know that a car is made of a few properties, including `make`,
`model`, `cc`, `year`. We can define a few rules like these:

    car = make
    car = make model
    car = make model 1.0
    car = model
    car = model 1.0

**Clear match by main attribute**

> 'I want a Ford'

We know it's a car because of the previous rule. This is simple to compute
(e.g 100% certainty, unless there are other meanings to _Ford_ leading to
ambiguity).

**Clear match by secondary attribute**

> 'I want a Fusion'

We know it's a car because of the previous rule. This is simple to compute
(e.g roughly 100% certainty, unless there are other meanings to _Fusion_
leading to ambiguity).

One thing that would help here is
[contextual score](https://github.com/kurko/nlu/issues/4):
the user could ultimately want a _Fusion [reactor]_, but we're in the cars
business so given context, it can only be a car unless the user says otherwise.

**Uncertain lookahead match**

> 'I want a Ford Abc'

We know `Ford` is a car, whereas `abc` is a potential model. Because `Ford` is
present, it adds to the score (e.g 75% certainty), and because we know the rule
to generalize, `car = make model`, we can assume that what comes next is a
_model_.

The caveat is that we assume it **could** be a model, but we don't know for
sure. So the model should output the assumption _and_ a uncertainty indication,
such as:

```ruby
{
  fn: :function,
  attrs: {
    make: 'Ford',
    model: 'Abc',
  },
  uncertain: {
    model: 'Abc'
  }
}
```

**Uncertain lookbehind match**

> 'I want a Abc Fusion'

We know Fusion is a _model_ and that before it generally comes the _make_. In
this case, `blabla` is unknown, so have a lower

```ruby
{
  fn: :function,
  attrs: {
    make: 'Abc',
    model: 'Fusion',
  },
  uncertain: {
    make: 'Abc'
  }
}
```

Connects to #5